fixing qot tick parameter.
git-svn-id: https://www.ibr.cs.tu-bs.de/svn/emma/trunk/src/ibrdtn@38360 1cffb75b-d306-6187-5d55-5a277715007a

+++ daemon/src/DTNTPWorker.h
@@ -61,7 +61,7 @@ namespace dtn
 			void sync(const TimeSyncMessage &msg, struct timeval &tv);
 
 			const dtn::daemon::Configuration::TimeSync &_conf;
-			size_t _qot_current_tic;
+			int _qot_current_tic;
 			double _sigma;
 			double _epsilon;
 			float _quality_diff;
